requirements.txt
PyQt6==6.6.1
PyQt6-Charts==6.6.0
SQLAlchemy==2.0.23
reportlab==4.0.7
python-docx==1.1.0
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
Pillow==10.1.0
python-dotenv==1.0.0
pytest==7.4.3
pywin32==306
